Support more than INT_MAX non-zero elements in sparse matrices.
This change allows ceres to solve larger sparse problems.

It is an intermediate step towards 64bit index support for sparse matrices (which would allow ceres to solve even larger problems and also allow for GPU acceleration in SuiteSparse).

This change does not switch to 64bit cell indices though, so the matrix size is still limited to INT_MAX elements per dimension.

However, values used to index into value and row/col arrays are promoted to 64-bit signed integers, so sparse matrices (except for CompressedRowSparseMatrix) can now have more than INT_MAX nonzero elements.

+++ internal/ceres/block_structure.h
@@ -64,7 +64,7 @@ struct Cell {
   // Column or row block id as the case maybe.
   int block_id;
   // Where in the values array of the jacobian is this cell located.
-  int position;
+  int64_t position;
 };
 
 // Order cell by their block_id;
